honeypot/network_monitor/requirements.txt
scapy==2.5.0
kafka-python==2.0.2
pyshark==0.6
yara-python==4.5.0
python-nmap==0.7.1
